feat: add more check, delete useless codes (#22)
* check if mint in runner

* add mut interface, check is all costed

* add height in context, check mint height

* delete useless code

* delete useless code

+++ vital-script/primitives/src/types/vrc20.rs
@@ -7,7 +7,7 @@ use super::MetaData;
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct VRC20MintMeta {
     pub mint_amount: u128,
-    pub mint_height: u64,
+    pub mint_height: u32,
     pub max_mints: u64,
 }
